RAG/backend/requirements.txt
flask
flask-cors
langchain
langchain-google-genai
langchain-huggingface
langchain-community
youtube-transcript-api
python-dotenv
